colision y por tipo de colision en punto
git-svn-id: https://svn.code.sf.net/p/warcom/code/trunk/Warcom_Game_Engine@78 bdfe1ff2-85b1-43ac-a073-3d879cfd3ae1

+++ engine.h
@@ -179,8 +179,8 @@ Local
 	float vY			= 0;     	//Velocidad Y
 	float fX			= 0;		//Posicion x coma flotante
 	float fY			= 0;		//Posicion y coma flotante
-	float alto			= 0;   		//Altura en pixeles del proceso
-	float ancho			= 0;   		//Ancho en pixeles del proceso
+	int   alto			= 0;   		//Altura en pixeles del proceso
+	int   ancho			= 0;   		//Ancho en pixeles del proceso
 	int   estado 		= 0;   		//Estado de la entidad
 	int	  numColPoints 	= 0;		//Numero de puntos de colision
 	struct colPoint[MAXCOLPOINTS] 		//Puntos deteccion colision
